refactor: split the integrations index page
Each rack and each creation form moves into its own component under
components/integrations/, leaving the page as the header plus the three
sections it composes. The exemption entry is gone from eslint.config.js.

Closes #985.
