+++ eslint.config.js
@@ -265,7 +265,6 @@ export default defineConfigWithVueTs(
             'resources/js/pages/teams/Edit.vue',
             'resources/js/pages/teams/Groups.vue',
             'resources/js/pages/teams/integrations/Bot.vue',
-            'resources/js/pages/teams/integrations/Index.vue',
         ],
         rules: {
             'max-lines': 'off',
